:lipstick: feat: Implement Project and Task Services
- Added ProjectService and ProjectServiceImpl for managing projects, including listing, creating, updating, archiving, and activating projects.
- Introduced CommentService for handling task comments with methods to add, update, and delete comments.
- Created CreateTaskService for task creation, ensuring responsible user validation and task assignment.
- Developed DeleteTaskService for task deletion, including moving tasks to a trashcan and unassigning from users.
- Implemented GetTaskService for retrieving tasks and counting tasks by status.
- Added TaskMaintenanceService for unassigning tasks from users based on team or project.
- Created UpdateTaskService for updating task details and managing user task assignments.
- Introduced LeaveTeamService for handling user removal from teams and associated project updates.
- Added GetUserService for retrieving user profiles and debugging user data.
- Implemented UpdateUserService for updating user information and reflecting changes across teams, projects, and tasks.
