+++ back-end/api-3/src/main/java/com/api_3/api_3/service/NotificationService.java
@@ -1,329 +1,60 @@
 package com.api_3.api_3.service;
 
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.api_3.api_3.model.entity.Notification;
-import com.api_3.api_3.model.entity.Notification.Type;
-import com.api_3.api_3.model.entity.Projects;
 import com.api_3.api_3.model.entity.Task;
-import com.api_3.api_3.model.entity.Teams;
-import com.api_3.api_3.model.entity.User;
-import com.api_3.api_3.repository.NotificationRepository;
-import com.api_3.api_3.repository.ProjectsRepository;
-import com.api_3.api_3.repository.TaskRepository;
-import com.api_3.api_3.repository.TeamsRepository;
-import com.api_3.api_3.repository.UserRepository;
 
 @Service
 public class NotificationService {
 
-    @Autowired private NotificationRepository notificationRepository;
-    @Autowired private ProjectsRepository projectsRepository;
-    @Autowired private TaskRepository taskRepository;
-    @Autowired private TeamsRepository teamsRepository;
-    @Autowired private UserRepository userRepository;
+    @Autowired private com.api_3.api_3.service.notification.TaskNotificationService taskNotificationService;
+    @Autowired private com.api_3.api_3.service.notification.ProjectNotificationService projectNotificationService;
+    @Autowired private com.api_3.api_3.service.notification.TeamNotificationService teamNotificationService;
+    @Autowired private com.api_3.api_3.service.notification.NotificationQueryService notificationQueryService;
+    @Autowired private com.api_3.api_3.service.notification.NotificationManagementService notificationManagementService;
 
-    // -------- Public API for task events --------
-    public void notifyTaskCreated(Task task) { notifyProjectMembers(task, Type.TASK_CREATED, "Tarefa criada", false); }
-    // Broadcast de atualização para todos os membros do projeto (exclui ator)
-    public void notifyTaskUpdatedBroadcast(Task task) { notifyProjectMembers(task, Type.TASK_UPDATED, "Tarefa atualizada", false); }
-    // Escopo: se houver responsável, notifica apenas ele; caso contrário, notifica todos do projeto (exclui ator)
-    public void notifyTaskUpdatedScoped(Task task) {
-        if (task.getResponsible() != null && task.getResponsible().uuid() != null) {
-            notifyOnlyUser(task, task.getResponsible().uuid(), Type.TASK_UPDATED, "Tarefa atualizada");
-        } else {
-            notifyProjectMembers(task, Type.TASK_UPDATED, "Tarefa atualizada", false);
-        }
-    }
-    // Escopo para exclusão: se houver responsável, notifica apenas ele; caso contrário, todos do projeto (exclui ator)
-    public void notifyTaskDeletedScoped(Task task) {
-        if (task.getResponsible() != null && task.getResponsible().uuid() != null) {
-            notifyOnlyUser(task, task.getResponsible().uuid(), Type.TASK_DELETED, "Tarefa excluída");
-        } else {
-            notifyProjectMembers(task, Type.TASK_DELETED, "Tarefa excluída", false);
-        }
-    }
+    public void notifyTaskCreated(Task task) { taskNotificationService.notifyTaskCreated(task); }
+    public void notifyTaskUpdatedBroadcast(Task task) { taskNotificationService.notifyTaskUpdatedBroadcast(task); }
+    public void notifyTaskUpdatedScoped(Task task) { taskNotificationService.notifyTaskUpdatedScoped(task); }
+    public void notifyTaskDeletedScoped(Task task) { taskNotificationService.notifyTaskDeletedScoped(task); }
 
-    // Assignment specific notifications (to the target user only)
-    public void notifyTaskAssigned(Task task, String newResponsibleUuid) {
-        if (newResponsibleUuid == null) return;
-        String actorUuid = getCurrentUserUuid();
-        if (actorUuid != null && actorUuid.equals(newResponsibleUuid)) return; // não notifica auto-atribuição
-        notifyOnlyUser(task, newResponsibleUuid, Type.TASK_ASSIGNED, "Você foi atribuído(a) à tarefa");
-    }
+    public void notifyTaskAssigned(Task task, String newResponsibleUuid) { taskNotificationService.notifyTaskAssigned(task, newResponsibleUuid); }
 
-    public void notifyTaskUnassigned(Task task, String oldResponsibleUuid) {
-        if (oldResponsibleUuid == null) return;
-        String actorUuid = getCurrentUserUuid();
-        if (actorUuid != null && actorUuid.equals(oldResponsibleUuid)) return; // não notifica auto-desatribuição
-        notifyOnlyUser(task, oldResponsibleUuid, Type.TASK_UNASSIGNED, "Você foi removido(a) da tarefa");
-    }
+    public void notifyTaskUnassigned(Task task, String oldResponsibleUuid) { taskNotificationService.notifyTaskUnassigned(task, oldResponsibleUuid); }
 
-    // Comentário: se houver responsável, notifica apenas ele (exceto se o autor for o próprio responsável);
-    // caso contrário, notifica todos os membros do projeto (exclui ator)
-    public void notifyTaskComment(Task task, String authorUuid) {
-        if (task.getResponsible() != null && task.getResponsible().uuid() != null) {
-            String r = task.getResponsible().uuid();
-            if (r.equals(authorUuid)) return; // evita notificar o próprio autor se ele for o responsável
-            notifyOnlyUser(task, r, Type.TASK_COMMENT, "Novo comentário na tarefa");
-        } else {
-            notifyProjectMembers(task, Type.TASK_COMMENT, "Novo comentário na tarefa", false);
-        }
-    }
+    public void notifyTaskComment(Task task, String authorUuid) { taskNotificationService.notifyTaskComment(task, authorUuid); }
 
-    // Due soon scanner runs hourly and warns the responsible 1 day before
-    @Scheduled(cron = "0 0 * * * *") // every hour
-    public void scanDueSoon() {
-        ZonedDateTime now = ZonedDateTime.now(ZoneId.systemDefault());
-        Date start = Date.from(now.toInstant());
-        Date end = Date.from(now.plusDays(1).toInstant());
-        List<Task> tasks = taskRepository.findByDueDateBetweenAndStatusNot(start, end, Task.Status.COMPLETED);
-        if (tasks == null || tasks.isEmpty()) return;
+    public void notifyTaskDueSoon(Task task) { taskNotificationService.notifyTaskDueSoon(task); }
 
-        // Notify only once per task per day
-        Date todayStart = Date.from(now.toLocalDate().atStartOfDay(now.getZone()).toInstant());
-
-        for (Task task : tasks) {
-            if (task.getResponsible() == null || task.getResponsible().uuid() == null) continue;
-            Optional<Notification> existing = notificationRepository.findFirstByTypeAndTaskUuidAndCreatedAtAfter(
-                Type.TASK_DUE_SOON, task.getUuid(), todayStart
-            );
-            if (existing.isPresent()) continue;
-
-            Notification n = base(task, Type.TASK_DUE_SOON, "Tarefa vence em 1 dia");
-            n.setUserUuid(task.getResponsible().uuid());
-            notificationRepository.save(n);
-        }
-    }
-
-    // -------- Queries and updates for controllers --------
     public List<Notification> listForCurrentUser(boolean unreadOnly) {
-        String userUuid = getCurrentUserUuid();
-        if (userUuid == null) return List.of();
-        return unreadOnly
-            ? notificationRepository.findByUserUuidAndReadFalseOrderByCreatedAtDesc(userUuid)
-            : notificationRepository.findByUserUuidOrderByCreatedAtDesc(userUuid);
+        return notificationQueryService.listForCurrentUser(unreadOnly);
     }
 
-    public Optional<Notification> markRead(String id) {
-        return notificationRepository.findById(id).map(n -> {
-            n.setRead(true);
-            n.setReadAt(new Date());
-            return notificationRepository.save(n);
-        });
-    }
+    public Optional<Notification> markRead(String id) { return notificationManagementService.markRead(id); }
 
-    public void delete(String id) { notificationRepository.deleteById(id); }
+    public void delete(String id) { notificationManagementService.delete(id); }
 
-    // -------- Internal helpers --------
-    // Notifica todos os membros de um projeto (exclui o ator se includeActor=false), com exclusões adicionais
-    private void notifyProjectMembers(Projects project, Type type, String message, boolean includeActor, java.util.Set<String> exclude) {
-        if (project == null || project.getMembers() == null || project.getMembers().isEmpty()) return;
-        String actorUuid = getCurrentUserUuid();
-        java.util.Set<String> recipients = new java.util.HashSet<>();
-        for (User.UserRef m : project.getMembers()) {
-            if (m == null || m.getUuid() == null) continue;
-            if (!includeActor && m.getUuid().equals(actorUuid)) continue;
-            if (exclude != null && exclude.contains(m.getUuid())) continue;
-            recipients.add(m.getUuid());
-        }
-        if (recipients.isEmpty()) return;
-        List<Notification> toSave = new ArrayList<>();
-        for (String dest : recipients) {
-            Notification n = new Notification();
-            n.setType(type);
-            n.setProjectUuid(project.getUuid());
-            n.setTeamUuid(project.getTeamUuid());
-            n.setMessage(message);
-            n.setCreatedAt(Date.from(Instant.now()));
-            n.setUserUuid(dest);
-            n.setActorUuid(actorUuid);
-            toSave.add(n);
-        }
-        if (!toSave.isEmpty()) notificationRepository.saveAll(toSave);
-    }
-
-    // Eventos de Projeto
     public void notifyProjectMemberJoined(String projectUuid, String joinedUserUuid) {
-        Projects project = projectsRepository.findById(projectUuid).orElse(null);
-        if (project == null) return;
-        String userName = userRepository.findById(joinedUserUuid).map(User::getName).orElse("Um usuário");
-        java.util.Set<String> exclude = new java.util.HashSet<>();
-        exclude.add(joinedUserUuid); // não notificar o próprio que entrou
-        // Notifica outros membros (exclui ator e o próprio que entrou)
-        notifyProjectMembers(project, Type.PROJECT_MEMBER_JOINED, userName + " entrou no projeto " + project.getName(), false, exclude);
-
-        // Fallback: se não houver destinatários além do ator e do próprio usuário que entrou,
-        // notifica o ator para registrar o evento (evita ficar sem registro no banco)
-        String actorUuid = getCurrentUserUuid();
-        boolean hasOtherRecipients = project.getMembers() != null && project.getMembers().stream()
-            .anyMatch(m -> m != null && m.getUuid() != null && !m.getUuid().equals(actorUuid) && !m.getUuid().equals(joinedUserUuid));
-        if (!hasOtherRecipients && actorUuid != null) {
-            Notification n = new Notification();
-            n.setType(Type.PROJECT_MEMBER_JOINED);
-            n.setProjectUuid(project.getUuid());
-            n.setTeamUuid(project.getTeamUuid());
-            n.setMessage(userName + " entrou no projeto " + project.getName());
-            n.setCreatedAt(Date.from(Instant.now()));
-            n.setUserUuid(actorUuid);
-            n.setActorUuid(actorUuid);
-            notificationRepository.save(n);
-        }
+        com.api_3.api_3.service.notification.ProjectNotificationService projectService = this.projectNotificationService;
+        if (projectService != null) projectService.notifyMemberJoined(projectUuid, joinedUserUuid);
     }
 
     public void notifyProjectMemberLeft(String projectUuid, String leftUserUuid) {
-        Projects project = projectsRepository.findById(projectUuid).orElse(null);
-        if (project == null) return;
-        String userName = userRepository.findById(leftUserUuid).map(User::getName).orElse("Um usuário");
-        java.util.Set<String> exclude = new java.util.HashSet<>();
-        exclude.add(leftUserUuid); // não notificar quem saiu
-        // Notifica outros membros (exclui ator e quem saiu)
-        notifyProjectMembers(project, Type.PROJECT_MEMBER_LEFT, userName + " saiu do projeto " + project.getName(), false, exclude);
-
-        // Fallback semelhante: se não houver outros destinatários, notifica o ator
-        String actorUuid = getCurrentUserUuid();
-        boolean hasOtherRecipients = project.getMembers() != null && project.getMembers().stream()
-            .anyMatch(m -> m != null && m.getUuid() != null && !m.getUuid().equals(actorUuid) && !m.getUuid().equals(leftUserUuid));
-        if (!hasOtherRecipients && actorUuid != null) {
-            Notification n = new Notification();
-            n.setType(Type.PROJECT_MEMBER_LEFT);
-            n.setProjectUuid(project.getUuid());
-            n.setTeamUuid(project.getTeamUuid());
-            n.setMessage(userName + " saiu do projeto " + project.getName());
-            n.setCreatedAt(Date.from(Instant.now()));
-            n.setUserUuid(actorUuid);
-            n.setActorUuid(actorUuid);
-            notificationRepository.save(n);
-        }
+        com.api_3.api_3.service.notification.ProjectNotificationService projectService = this.projectNotificationService;
+        if (projectService != null) projectService.notifyMemberLeft(projectUuid, leftUserUuid);
     }
-    // Notifica todos os membros de um time (exclui o ator se includeActor=false), com exclusões adicionais
-    private void notifyTeamMembers(Teams team, Type type, String message, boolean includeActor, java.util.Set<String> exclude) {
-        if (team == null || team.getMembers() == null || team.getMembers().isEmpty()) return;
-        String actorUuid = getCurrentUserUuid();
-        java.util.Set<String> recipients = new java.util.HashSet<>();
-        for (User.UserRef m : team.getMembers()) {
-            if (m == null || m.getUuid() == null) continue;
-            if (!includeActor && m.getUuid().equals(actorUuid)) continue;
-            if (exclude != null && exclude.contains(m.getUuid())) continue;
-            recipients.add(m.getUuid());
-        }
-        if (recipients.isEmpty()) return;
-        List<Notification> toSave = new ArrayList<>();
-        for (String dest : recipients) {
-            Notification n = new Notification();
-            n.setType(type);
-            n.setTeamUuid(team.getUuid());
-            n.setMessage(message);
-            n.setCreatedAt(Date.from(Instant.now()));
-            n.setUserUuid(dest);
-            n.setActorUuid(actorUuid);
-            toSave.add(n);
-        }
-        if (!toSave.isEmpty()) notificationRepository.saveAll(toSave);
-    }
-
-    // Eventos de Team
     public void notifyTeamMemberJoined(String teamUuid, String joinedUserUuid) {
-        Teams team = teamsRepository.findById(teamUuid).orElse(null);
-        if (team == null) return;
-        String userName = userRepository.findById(joinedUserUuid).map(User::getName).orElse("Um usuário");
-        java.util.Set<String> exclude = new java.util.HashSet<>();
-        exclude.add(joinedUserUuid); // não notificar o próprio que entrou
-        notifyTeamMembers(team, Type.TEAM_MEMBER_JOINED, userName + " entrou na equipe " + team.getName(), false, exclude);
+        com.api_3.api_3.service.notification.TeamNotificationService teamService = this.teamNotificationService;
+        if (teamService != null) teamService.notifyMemberJoined(teamUuid, joinedUserUuid);
     }
 
     public void notifyTeamMemberLeft(String teamUuid, String leftUserUuid) {
-        Teams team = teamsRepository.findById(teamUuid).orElse(null);
-        if (team == null) return;
-        String userName = userRepository.findById(leftUserUuid).map(User::getName).orElse("Um usuário");
-        java.util.Set<String> exclude = new java.util.HashSet<>();
-        exclude.add(leftUserUuid); // não notificar quem saiu
-        notifyTeamMembers(team, Type.TEAM_MEMBER_LEFT, userName + " saiu da equipe " + team.getName(), false, exclude);
-    }
-
-    private void notifyProjectMembers(Task task, Type type, String defaultMessage, boolean includeActor) {
-        if (task.getProjectUuid() == null) return;
-        Projects project = projectsRepository.findById(task.getProjectUuid()).orElse(null);
-        if (project == null || project.getMembers() == null || project.getMembers().isEmpty()) return;
-        String actorUuid = getCurrentUserUuid();
-        Set<String> recipients = project.getMembers().stream()
-            .map(User.UserRef::uuid)
-            .filter(u -> u != null && (includeActor || !u.equals(actorUuid)))
-            .collect(Collectors.toCollection(HashSet::new));
-
-        // Also ensure responsible is notified (if present and not the actor)
-        if (task.getResponsible() != null && task.getResponsible().uuid() != null) {
-            String r = task.getResponsible().uuid();
-            if (!r.equals(actorUuid)) recipients.add(r);
-        }
-
-        if (recipients.isEmpty()) return;
-
-        createAndSaveNotifications(task, type, defaultMessage, recipients, actorUuid);
-    }
-
-    private Notification base(Task task, Type type, String message) {
-        Notification n = new Notification();
-        n.setType(type);
-        n.setTaskUuid(task.getUuid());
-        n.setTaskTitle(task.getTitle());
-        n.setProjectUuid(task.getProjectUuid());
-        n.setTeamUuid(task.getEquip_uuid());
-        n.setMessage(message);
-        n.setCreatedAt(Date.from(Instant.now()));
-        return n;
-    }
-
-    private void notifyOnlyUser(Task task, String userUuid, Type type, String message) {
-        Notification n = base(task, type, message);
-        n.setUserUuid(userUuid);
-        n.setActorUuid(getCurrentUserUuid());
-        notificationRepository.save(n);
-    }
-
-    private void createAndSaveNotifications(Task task, Type type, String message, Set<String> recipients, String actorUuid) {
-        List<Notification> toSave = new ArrayList<>();
-        for (String dest : recipients) {
-            Notification n = base(task, type, message);
-            n.setUserUuid(dest);
-            n.setActorUuid(actorUuid);
-            toSave.add(n);
-        }
-        if (!toSave.isEmpty()) notificationRepository.saveAll(toSave);
-    }
-
-    private String getCurrentUserUuid() {
-        try {
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-            if (auth == null) return null;
-            String username = null;
-            Object principal = auth.getPrincipal();
-            if (principal instanceof org.springframework.security.core.userdetails.UserDetails ud) {
-                username = ud.getUsername();
-            } else {
-                username = auth.getName();
-            }
-            if (username == null) return null;
-            return userRepository.findByEmailIgnoreCase(username)
-                    .map(User::getUuid)
-                    .orElse(null);
-        } catch (Exception e) {
-            return null;
-        }
+        com.api_3.api_3.service.notification.TeamNotificationService teamService = this.teamNotificationService;
+        if (teamService != null) teamService.notifyMemberLeft(teamUuid, leftUserUuid);
     }
 }
