+++ back-end/api-3/src/main/java/com/api_3/api_3/service/project/LeaveProjectService.java
@@ -1,4 +1,4 @@
-package com.api_3.api_3.service;
+package com.api_3.api_3.service.project;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -14,6 +14,8 @@ import com.api_3.api_3.repository.ProjectsRepository;
 import com.api_3.api_3.repository.TaskRepository;
 import com.api_3.api_3.repository.TeamsRepository;
 import com.api_3.api_3.repository.UserRepository;
+import com.api_3.api_3.service.NotificationService;
+import com.api_3.api_3.service.task.TaskMaintenanceService;
 
 @Service
 public class LeaveProjectService {
@@ -41,17 +43,14 @@ public class LeaveProjectService {
         Projects project = projectsRepository.findById(projectUuid)
                 .orElseThrow(() -> new ProjectNotFoundException("Project não encontrado com o ID: " + projectUuid));
 
-        // validar se o utilizador é membro do projeto
         boolean isMember = project.getMembers().stream().anyMatch(m -> m.getUuid().equals(userUuid));
         if (!isMember) {
             throw new SecurityException("Acesso negado. O utilizador não é membro deste projeto.");
         }
 
-        // remover utilizador dos membros do projeto
         project.getMembers().removeIf(member -> member.getUuid().equals(userUuid));
     projectsRepository.save(project);
 
-        // buscar apenas as tasks deste projeto que têm o user como responsible
         List<Task> tasks = taskRepository.findByProjectUuidAndResponsibleUuid(projectUuid, userUuid);
         if (tasks != null && !tasks.isEmpty()) {
             for (Task t : tasks) {
@@ -60,7 +59,6 @@ public class LeaveProjectService {
             taskRepository.saveAll(tasks);
         }
 
-        // remover referências dessas tasks na lista de tasks do usuário (se existir)
         userRepository.findById(userUuid).ifPresent(user -> {
             if (user.getTasks() != null) {
                 user.getTasks().removeIf(tRef -> projectUuid.equals(tRef.getProjectUuid()));
@@ -68,12 +66,8 @@ public class LeaveProjectService {
             userRepository.save(user);
         });
 
-        // desatribuir tarefas relacionadas com este projeto para este utilizador
-        // (opcional/redundante)
         taskMaintenanceService.unassignForProject(projectUuid, userUuid);
 
-        // se o projeto ficar sem membros, remover o projeto e remover referência no
-        // Team; caso contrário, notificar os membros que o usuário saiu
         if (project.getMembers() == null || project.getMembers().isEmpty()) {
             projectsRepository.deleteById(projectUuid);
 
@@ -87,7 +81,6 @@ public class LeaveProjectService {
                 });
             }
         } else {
-            // Notificar membros restantes do projeto sobre a saída
             notificationService.notifyProjectMemberLeft(projectUuid, userUuid);
         }
     }
