+++ back-end/api-3/src/main/java/com/api_3/api_3/service/project/ProjectService.java
@@ -1,4 +1,4 @@
-package com.api_3.api_3.service;
+package com.api_3.api_3.service.project;
 
 import java.util.List;
