+++ back-end/api-3/src/main/java/com/api_3/api_3/service/task/GetTaskService.java
@@ -1,13 +1,14 @@
-package com.api_3.api_3.service;
+package com.api_3.api_3.service.task;
+
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.api_3.api_3.exception.TaskNotFoundException;
 import com.api_3.api_3.model.entity.Task;
 import com.api_3.api_3.repository.TaskRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
 
 @Service
 public class GetTaskService {
