+++ back-end/api-3/src/main/java/com/api_3/api_3/service/team/LeaveTeamService.java
@@ -1,4 +1,4 @@
-package com.api_3.api_3.service;
+package com.api_3.api_3.service.team;
 
 import java.util.List;
 
@@ -14,6 +14,8 @@ import com.api_3.api_3.model.entity.User;
 import com.api_3.api_3.repository.ProjectsRepository;
 import com.api_3.api_3.repository.TeamsRepository;
 import com.api_3.api_3.repository.UserRepository;
+import com.api_3.api_3.service.NotificationService;
+import com.api_3.api_3.service.task.TaskMaintenanceService;
 
 @Service
 public class LeaveTeamService {
@@ -28,7 +30,7 @@ public class LeaveTeamService {
     private ProjectsRepository projectsRepository;
 
     @Autowired
-    private TaskMaintenanceService taskMaintenanceService; // Injetamos o serviço existente
+    private TaskMaintenanceService taskMaintenanceService;
 
     @Autowired
     private NotificationService notificationService;
@@ -41,37 +43,28 @@ public class LeaveTeamService {
         User user = userRepository.findById(userUuid)
                 .orElseThrow(() -> new UserNotFoundException("Utilizador não encontrado com o ID: " + userUuid));
 
-        //  Validar se o utilizador é membro da equipa
         boolean isMember = team.getMembers().stream().anyMatch(m -> m.getUuid().equals(userUuid));
         if (!isMember) {
             throw new SecurityException("Acesso negado. O utilizador não é membro desta equipa.");
         }
 
-        // Remover o utilizador da lista de membros da equipe
         team.getMembers().removeIf(member -> member.getUuid().equals(userUuid));
         teamsRepository.save(team);
 
-        // Remover a equipe da lista de equipas do utilizador
         user.getTeams().removeIf(teamRef -> teamRef.getUuid().equals(teamUuid));
         userRepository.save(user);
 
-        // Remover o utilizador de todos os projetos associados à equipe (NOVA LÓGICA)
         List<Projects> projects = projectsRepository.findByTeamUuid(teamUuid);
         for (Projects project : projects) {
             project.getMembers().removeIf(member -> member.getUuid().equals(userUuid));
             projectsRepository.save(project);
         }
 
-        // Chamar o serviço de manutenção para desatribuir as tarefas (REUTILIZAÇÃO)
         taskMaintenanceService.unassignForTeam(teamUuid, userUuid);
 
-        // Se a equipe ficar sem membros, removê-la do banco
         if (team.getMembers() == null || team.getMembers().isEmpty()) {
             teamsRepository.deleteById(teamUuid);
-            // Opcional: remover/arquivar projetos e tarefas associados ao time
-            // Mantido simples conforme requisito: apenas excluir o Team quando sem participantes
         } else {
-            // Notificar os membros restantes sobre a saída
             notificationService.notifyTeamMemberLeft(teamUuid, userUuid);
         }
     }
