+++ back-end/api-3/src/main/java/com/api_3/api_3/service/user/UpdateUserService.java
@@ -1,23 +1,23 @@
-package com.api_3.api_3.service;
+package com.api_3.api_3.service.user;
 
-import org.springframework.beans.factory.annotation.Autowired; 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.List;
-import java.util.ArrayList; 
-import java.util.stream.Collectors; 
 
 import com.api_3.api_3.dto.request.UpdateUserRequest;
-import com.api_3.api_3.model.entity.User;
-import com.api_3.api_3.model.entity.Teams; 
 import com.api_3.api_3.model.entity.Projects;
-import com.api_3.api_3.model.entity.Task; 
-import com.api_3.api_3.repository.UserRepository;
-import com.api_3.api_3.repository.TeamsRepository; 
+import com.api_3.api_3.model.entity.Task;
+import com.api_3.api_3.model.entity.Teams;
+import com.api_3.api_3.model.entity.User; 
 import com.api_3.api_3.repository.ProjectsRepository;
 import com.api_3.api_3.repository.TaskRepository; 
+import com.api_3.api_3.repository.TeamsRepository;
+import com.api_3.api_3.repository.UserRepository; 
 
 
 @Service
@@ -28,7 +28,6 @@ public class UpdateUserService {
     private final ProjectsRepository projectsRepository; 
     private final TaskRepository taskRepository; 
 
-    @Autowired
     public UpdateUserService(UserRepository userRepository,
                              TeamsRepository teamsRepository,
                              ProjectsRepository projectsRepository,
@@ -60,15 +59,12 @@ public class UpdateUserService {
             user.setImg(request.getImg().trim());
             imgChanged = true;
         }
-        // Se nada mudou, retorna o usuário sem salvar ou cascatear
         if (!nameChanged && !imgChanged) {
             return user;
         }
 
-        // Salva a atualização principal no documento do usuário
         User updatedUser = userRepository.save(user);
 
-        // Dispara as atualizações em cascata SE algo mudou
         updateUserReferences(updatedUser);
 
         return updatedUser;
@@ -78,11 +74,9 @@ public class UpdateUserService {
         String userId = updatedUser.getUuid();
         User.UserRef updatedUserRef = updatedUser.toRef(); 
 
-        // Atualizar em Teams
         List<Teams> teamsToUpdate = teamsRepository.findByMembersUuid(userId);
         List<Teams> modifiedTeams = new ArrayList<>();
         for (Teams team : teamsToUpdate) {
-            boolean teamModified = false;
             List<User.UserRef> updatedMembers = team.getMembers().stream()
                 .map(member -> {
                     if (userId.equals(member.uuid())) {
@@ -93,7 +87,6 @@ public class UpdateUserService {
                 })
                 .collect(Collectors.toList());
 
-            // Verifica se a lista realmente mudou para evitar salvamentos desnecessários
             if (!updatedMembers.equals(team.getMembers())) {
                  team.setMembers(updatedMembers);
                  modifiedTeams.add(team);
@@ -104,7 +97,6 @@ public class UpdateUserService {
         }
 
 
-        // Atualizar em Projects
         List<Projects> projectsToUpdate = projectsRepository.findByMembersUuid(userId);
          List<Projects> modifiedProjects = new ArrayList<>();
         for (Projects project : projectsToUpdate) {
@@ -121,7 +113,6 @@ public class UpdateUserService {
             projectsRepository.saveAll(modifiedProjects);
         }
 
-        //  Atualizar em Tasks 
         List<Task> tasksToUpdate = taskRepository.findByResponsibleUuid(userId);
          List<Task> modifiedTasks = new ArrayList<>();
         for (Task task : tasksToUpdate) {
